making: up, schema and models matching

+++ migrations/2025-05-12-180906_create_products/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE products(
    Id_products INTEGER,
    description TEXT,
    name TEXT,
-   price NUMERIC,
+   price REAL,
    image_url TEXT,
    PRIMARY KEY(Id_products)
 );
